+++ lib/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- remove chart `onInit` in favor of chart API refs on `Chart`, `OptionsChart`, `YieldCurveChart`, and `CustomChart`, and add `containerRef` for accessing the wrapper div element
 
 ## [2.5.0] - 2026-06-24
 ### Fixed